refactor(stigma): Move StigmaDetails do domain package.
