+++ stigma/src/main/java/io/nixer/nixerplugin/stigma/evaluate/StigmaService.java
@@ -6,7 +6,7 @@ import javax.annotation.Nullable;
 import com.google.common.base.Preconditions;
 import io.nixer.nixerplugin.stigma.domain.Stigma;
 import io.nixer.nixerplugin.stigma.domain.StigmaStatus;
-import io.nixer.nixerplugin.stigma.storage.StigmaDetails;
+import io.nixer.nixerplugin.stigma.domain.StigmaDetails;
 import io.nixer.nixerplugin.stigma.storage.StigmaStorage;
 import io.nixer.nixerplugin.stigma.generate.StigmaGenerator;
 import org.slf4j.Logger;
