+++ stigma/src/main/java/io/nixer/nixerplugin/stigma/storage/StigmaStorage.java
@@ -4,6 +4,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 import io.nixer.nixerplugin.stigma.domain.Stigma;
+import io.nixer.nixerplugin.stigma.domain.StigmaDetails;
 import io.nixer.nixerplugin.stigma.domain.StigmaStatus;
 
 /**
