+++ stigma/src/test/java/io/nixer/nixerplugin/stigma/evaluate/StigmaActionEvaluatorTest.java
@@ -4,7 +4,7 @@ import java.time.Instant;
 
 import io.nixer.nixerplugin.stigma.domain.RawStigmaToken;
 import io.nixer.nixerplugin.stigma.domain.Stigma;
-import io.nixer.nixerplugin.stigma.storage.StigmaDetails;
+import io.nixer.nixerplugin.stigma.domain.StigmaDetails;
 import io.nixer.nixerplugin.stigma.token.read.StigmaExtractor;
 import io.nixer.nixerplugin.stigma.token.create.StigmaTokenFactory;
 import org.junit.jupiter.api.BeforeEach;
