+++ stigma/src/test/java/io/nixer/nixerplugin/stigma/evaluate/StigmaServiceTest.java
@@ -4,7 +4,7 @@ import java.time.Instant;
 
 import io.nixer.nixerplugin.stigma.domain.Stigma;
 import io.nixer.nixerplugin.stigma.domain.StigmaStatus;
-import io.nixer.nixerplugin.stigma.storage.StigmaDetails;
+import io.nixer.nixerplugin.stigma.domain.StigmaDetails;
 import io.nixer.nixerplugin.stigma.storage.StigmaStorage;
 import io.nixer.nixerplugin.stigma.generate.StigmaGenerator;
 import org.junit.jupiter.api.Test;
